pyproject.toml
[tool.poetry]
name = "Aoi"
version = "0.1.0"
description = "Aoi is Discord bot to grant role"
authors = ["Yoiduki <90896806+y-muen@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {extras = ["voice"], version = "^1.7.3"}
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
